+++ client/src/App.js
@@ -2,7 +2,7 @@
 import React, { Component } from 'react';
 import { Route, Switch } from 'react-router-dom';
 import './App.css';
-import Home from './components/Home';
+import Admin from './components/Admin';
 import SignUp from './components/SignUp';
 import LogIn from './components/LogIn';
 import List from './components/List';
@@ -15,7 +15,7 @@ class App extends Component {
     const App = () => (
       <div>
         <Switch>
-          <Route exact path='/' component={Home}/>
+          <Route exact path='/admin' component={Admin}/>
           <Route path='/list' component={List}/>
           <Route exact path='/login' component={LogIn}/>
           <Route path='/signup' component={SignUp}/>
